mv map_reg
